+++ app/src/main/java/com/heyzeusv/plutuswallet/ui/AlertDialogComposables.kt
@@ -1,4 +1,4 @@
-package com.heyzeusv.plutuswallet.util
+package com.heyzeusv.plutuswallet.ui
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -38,7 +38,6 @@ import androidx.compose.ui.window.Dialog
 import com.heyzeusv.plutuswallet.R
 import com.heyzeusv.plutuswallet.data.model.Account
 import com.heyzeusv.plutuswallet.data.model.DataInterface
-import com.heyzeusv.plutuswallet.ui.PreviewHelper
 import com.heyzeusv.plutuswallet.util.theme.LocalPWColors
 import com.heyzeusv.plutuswallet.util.theme.alertDialogButton
