+++ app/src/main/java/com/heyzeusv/plutuswallet/ui/settings/SettingsComposables.kt
@@ -29,10 +29,10 @@ import androidx.compose.ui.unit.sp
 import androidx.preference.PreferenceManager
 import com.heyzeusv.plutuswallet.R
 import com.heyzeusv.plutuswallet.ui.PreviewHelper
+import com.heyzeusv.plutuswallet.ui.PWListAlertDialog
+import com.heyzeusv.plutuswallet.ui.PWAlertDialog
 import com.heyzeusv.plutuswallet.util.AppBarActions
 import com.heyzeusv.plutuswallet.util.Key
-import com.heyzeusv.plutuswallet.util.PWListAlertDialog
-import com.heyzeusv.plutuswallet.util.PWAlertDialog
 import com.heyzeusv.plutuswallet.util.PreferenceHelper.get
 import com.heyzeusv.plutuswallet.util.PreferenceHelper.set
 import com.heyzeusv.plutuswallet.util.SettingOptions
